+++ app/src/main/java/io/github/sds100/keymapper/actions/ActionUiHelper.kt
@@ -223,6 +223,8 @@ class ActionUiHelper(
                         is ActionData.ControlMediaForApp.PreviousTrack -> R.string.action_previous_track_package_formatted
                         is ActionData.ControlMediaForApp.Rewind -> R.string.action_rewind_package_formatted
                         is ActionData.ControlMediaForApp.Stop -> R.string.action_stop_media_package_formatted
+                        is ActionData.ControlMediaForApp.StepForward -> R.string.action_step_forward_media_package_formatted
+                        is ActionData.ControlMediaForApp.StepBackward -> R.string.action_step_backward_media_package_formatted
                     }
 
                     getString(resId, appName)
@@ -237,6 +239,8 @@ class ActionUiHelper(
                         is ActionData.ControlMediaForApp.PreviousTrack -> R.string.action_previous_track_package
                         is ActionData.ControlMediaForApp.Rewind -> R.string.action_rewind_package
                         is ActionData.ControlMediaForApp.Stop -> R.string.action_stop_media_package
+                        is ActionData.ControlMediaForApp.StepForward -> R.string.action_step_forward_media_package
+                        is ActionData.ControlMediaForApp.StepBackward -> R.string.action_step_backward_media_package
                     }
 
                     getString(resId)
@@ -456,6 +460,8 @@ class ActionUiHelper(
         ActionData.ControlMedia.PreviousTrack -> getString(R.string.action_previous_track)
         ActionData.ControlMedia.Rewind -> getString(R.string.action_rewind)
         ActionData.ControlMedia.Stop -> getString(R.string.action_stop_media)
+        ActionData.ControlMedia.StepForward -> getString(R.string.action_step_forward_media)
+        ActionData.ControlMedia.StepBackward -> getString(R.string.action_step_backward_media)
 
         ActionData.CopyText -> getString(R.string.action_text_copy)
         ActionData.CutText -> getString(R.string.action_text_cut)