+++ app/src/main/java/io/github/sds100/keymapper/actions/CreateActionDelegate.kt
@@ -182,6 +182,8 @@ class CreateActionDelegate(
             ActionId.FAST_FORWARD_PACKAGE,
             ActionId.REWIND_PACKAGE,
             ActionId.STOP_MEDIA_PACKAGE,
+            ActionId.STEP_FORWARD_PACKAGE,
+            ActionId.STEP_BACKWARD_PACKAGE,
             -> {
                 val packageName =
                     navigate(
@@ -214,6 +216,12 @@ class CreateActionDelegate(
 
                     ActionId.STOP_MEDIA_PACKAGE ->
                         ActionData.ControlMediaForApp.Stop(packageName)
+
+                    ActionId.STEP_FORWARD_PACKAGE ->
+                        ActionData.ControlMediaForApp.StepForward(packageName)
+
+                    ActionId.STEP_BACKWARD_PACKAGE ->
+                        ActionData.ControlMediaForApp.StepBackward(packageName)
 
                     else -> throw Exception("don't know how to create action for $actionId")
                 }
@@ -734,6 +742,8 @@ class CreateActionDelegate(
             ActionId.FAST_FORWARD -> return ActionData.ControlMedia.FastForward
             ActionId.REWIND -> return ActionData.ControlMedia.Rewind
             ActionId.STOP_MEDIA -> return ActionData.ControlMedia.Stop
+            ActionId.STEP_FORWARD -> return ActionData.ControlMedia.StepForward
+            ActionId.STEP_BACKWARD -> return ActionData.ControlMedia.StepBackward
 
             ActionId.GO_BACK -> return ActionData.GoBack
             ActionId.GO_HOME -> return ActionData.GoHome