+++ app/src/main/java/io/github/sds100/keymapper/actions/PerformActionsUseCase.kt
@@ -227,6 +227,14 @@ class PerformActionsUseCaseImpl(
                 result = mediaAdapter.stop(action.packageName)
             }
 
+            is ActionData.ControlMediaForApp.StepForward -> {
+                result = mediaAdapter.stepForward(action.packageName)
+            }
+
+            is ActionData.ControlMediaForApp.StepBackward -> {
+                result = mediaAdapter.stepBackward(action.packageName)
+            }
+
             is ActionData.Rotation.CycleRotations -> {
                 result = displayAdapter.disableAutoRotate().then {
                     val currentOrientation = displayAdapter.cachedOrientation
@@ -553,6 +561,14 @@ class PerformActionsUseCaseImpl(
 
             is ActionData.ControlMedia.Stop -> {
                 result = mediaAdapter.stop()
+            }
+
+            is ActionData.ControlMedia.StepForward -> {
+                result = mediaAdapter.stepForward()
+            }
+
+            is ActionData.ControlMedia.StepBackward -> {
+                result = mediaAdapter.stepBackward()
             }
 
             is ActionData.GoBack -> {