+++ app/src/main/java/io/github/sds100/keymapper/system/media/MediaAdapter.kt
@@ -36,6 +36,8 @@ interface MediaAdapter {
     fun previousTrack(packageName: String? = null): Result<*>
     fun nextTrack(packageName: String? = null): Result<*>
     fun stop(packageName: String? = null): Result<*>
+    fun stepForward(packageName: String? = null): Result<*>
+    fun stepBackward(packageName: String? = null): Result<*>
 
     fun playFile(uri: String, stream: VolumeStream): Result<*>
     fun stopFileMedia(): Result<*>